Nested for loop is not the right way to go
After some time figuring out how this is working, I came to the realization that we are iterating through each `_value` and then iterating through those values again. This ends up being a O(n^2) complexity.

It is better to place all the items into the widget list first, and then check the existence of the item in that list and select it. This makes the function a O(2n) complexity this will take significantly less time to compute