Refactoring getChangedValues
This new improved version cuts down on logic. Instead all we need to do is reference the algorithm that we are curious about, and everything else will do the extra work.

The reason why this works is because, all these algorithms accept the dictionary's keys as input. If the keys are wrong, the algorithm breaks. Thus we can safely assume that using the dictionaries keys to check for changes on both dict input and alg input will work