Version 3.3.0 (#158)
* Update travis

* Move the exception logger to it's own function

Simplify and reduce

* have stdout set to the console window

This way we are not in the dark about what is actually happening beneath.

Exceptions will still occur naturally as always, and the logger can take over during bad Exceptions

* Update version

Update the version to reflect the new changes.

* Add a new lines

This will display documentation better.

* Add documentation

* Simplify

Make it more clear what is going on.

* adding isGuiChanged

* PEP8

* Add the ability to detect state changes in UI

* We no longer need checkoptions

As per previous commit, state changes need to be detectede across all UI, this will fix this problem

* Fix #127

Give proper versioning in about section

* Add version recognition to logger

* ReAdding Qtickle temporarily

* Rename, bad practive to use parents name

* making restore filename available for use

* Adding selective restore function

* Still trying to figure out how to get this Widget to show up

* PEP 8

* Still trying to figure out how to get this Widget to show

* Rename, bad practive to use parents name

* Adding notes and selective ui functionality

* Revert "Making modules handle changed input better"

This reverts commit 8a15f7acde4e6223d855070cecb8750dfa5527a9.

Revert back, as we will be doing a better approach

* Add functionality for guiState

* Documentation

* About finally appears correctly

* PEP8

* Maybe fixed #134

Update some of the documentation
This new functionality has not been fully tested. This should be done to ensure that no weird problems emerge fromt his.

* Check options is still haunting us like the plague

* Simplify and delete, also better naming conventions

* Adding selectiveRestore to overloaded modules

* Nested for loop is not the right way to go

After some time figuring out how this is working, I came to the realization that we are iterating through each `_value` and then iterating through those values again. This ends up being a O(n^2) complexity.

It is better to place all the items into the widget list first, and then check the existence of the item in that list and select it. This makes the function a O(2n) complexity this will take significantly less time to compute

* Spelling

* Temporarily remove selectiveRestore

* Remove float and int

The .value() gives us the proper type. Casting the type again is redundant but may cause problems if we didn't cast correctly.

* Utilize the changed names

* Changed values function

* Refactoring getChangedValues

This new improved version cuts down on logic. Instead all we need to do is reference the algorithm that we are curious about, and everything else will do the extra work.

The reason why this works is because, all these algorithms accept the dictionary's keys as input. If the keys are wrong, the algorithm breaks. Thus we can safely assume that using the dictionaries keys to check for changes on both dict input and alg input will work

* Do symmetric checking of methodParameters

We can now check our parameters for changes.

* Force the comboBox into a particular size

* We can now have DataBox start independently

I've also taken away the multithreading from DataBox. The reason for this has to do with the fact that Databox is now on it's own Form Thread. In other words we do not have to worry about it affecting the main GUI when we are trying to load some data, thus it is alright that we don't utilize multithreading here.

* Use raise exception in order to stop the UI

but not grey out the UI.

* Updating PandasModel to be more fast

* Update Version

* Remove Qtickle for now